pyproject.toml
[project]
name = "MusicDownloader"
version = "0.1.0"
description = "Add your description here"  # TODO
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dacite>=1.9.2",
    "line-profiler-pycharm>=1.2.0",
    "mutagen>=1.47.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pyright>=1.1.400",
    "pyside6>=6.9.0",
    "python-vlc>=3.0.21203",
    "soundfile>=0.13.1",
    "tqdm>=4.67.1",
]

[tool.ruff]
src = ["src"]
line-length = 120
target-version = "py313"
fix = true

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pydocstyle]
convention = "numpy"
